refactor(cli): converge cmd/pigo/main.go to thin entry (#369)
Consolidate cliOptions, applyFileConfig, and dispatch into a single
319-line main.go and delete run.go/config_file.go. Merge the three
package-main test files into one main_test.go. cmd/pigo now contains
only main.go and main_test.go; all run/REPL/headless/config logic lives
under internal/cli/*.
