+++ docs/issue#0369.html
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Implementation Notes — Issue #0369</title>
+  <style>
+    * { margin: 0; padding: 0; box-sizing: border-box; }
+    body {
+      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+      background: #FAF9F6;
+      color: #1a1a1a;
+      padding: 2.5rem 2rem;
+      line-height: 1.7;
+    }
+    .container { max-width: 800px; margin: 0 auto; }
+    h1 { font-size: 1.375rem; font-weight: 700; margin-bottom: 0.25rem; }
+    .meta { color: #8B8680; font-size: 0.8125rem; margin-bottom: 2rem; }
+    h2 {
+      font-size: 1rem;
+      font-weight: 600;
+      margin-top: 2rem;
+      margin-bottom: 0.75rem;
+      padding-bottom: 0.375rem;
+      border-bottom: 1px solid #E8E4DE;
+      display: flex; align-items: center; gap: 0.5rem;
+    }
+    .dot {
+      width: 8px; height: 8px; border-radius: 50%; display: inline-block;
+    }
+    .dot.design { background: #5B8A72; }
+    .dot.deviation { background: #D97757; }
+    .dot.tradeoff { background: #4A6FA5; }
+    .dot.question { background: #D4A843; }
+    .item {
+      background: #FFFFFF;
+      border: 1px solid #E8E4DE;
+      border-radius: 8px;
+      padding: 1rem 1.25rem;
+      margin-bottom: 0.75rem;
+      box-shadow: 0 1px 3px rgba(0,0,0,0.03);
+    }
+    .item h3 { font-size: 0.875rem; font-weight: 600; margin-bottom: 0.375rem; }
+    .item p { font-size: 0.8125rem; color: #4A4540; margin-bottom: 0.375rem; }
+    .label {
+      display: inline-block;
+      font-size: 0.6875rem;
+      font-weight: 500;
+      padding: 0.125rem 0.5rem;
+      border-radius: 4px;
+      margin-right: 0.375rem;
+    }
+    .label-design { background: #F5F8F6; color: #5B8A72; border: 1px solid #5B8A72; }
+    .label-deviation { background: #FFF5F0; color: #D97757; border: 1px solid #D97757; }
+    .label-tradeoff { background: #F0F4F8; color: #4A6FA5; border: 1px solid #4A6FA5; }
+    .label-question { background: #FDF8F0; color: #D4A843; border: 1px solid #D4A843; }
+    .none { color: #B0AAA4; font-style: italic; font-size: 0.8125rem; }
+    .footer {
+      text-align: center; margin-top: 2.5rem; color: #B0AAA4;
+      font-size: 0.6875rem;
+    }
+  </style>
+</head>
+<body>
+  <div class="container">
+    <h1>Implementation Notes</h1>
+    <p class="meta">Issue <a href="https://github.com/smallnest/pigo/issues/369">#369</a> &mdash; 收敛 cmd/pigo/main.go 为薄入口 (US-014) &mdash; 2026-07-28</p>
+
+    <h2><span class="dot design"></span> Design Decisions</h2>
+    <div class="item">
+      <h3><span class="label label-design">Decision</span> Consolidate all package-main source into a single main.go</h3>
+      <p><strong>Ambiguity:</strong> The AC says main.go should hold only main()/cliOptions/flag defs/dispatch/version, and "cmd/pigo 除 main.go 外无其他源文件" — but dispatch lived in run.go and applyFileConfig in config_file.go.</p>
+      <p><strong>Choice:</strong> Moved <code>cliOptions</code>, <code>dispatch</code>, and <code>applyFileConfig</code> verbatim into main.go and deleted run.go/config_file.go. main.go is now 319 lines (&lt; ~400 budget), containing exactly: version metadata, cliOptions, main(), applyFileConfig, dispatch.</p>
+      <p><strong>Rationale:</strong> These three are all CLI-glue (flag parse → config overlay → mode dispatch) that must stay in package main after every subpackage migration. Co-locating them makes main.go the single, self-contained entry point the AC calls for; the real logic already lives in internal/cli/*.</p>
+    </div>
+
+    <h2><span class="dot deviation"></span> Deviations</h2>
+    <div class="item">
+      <h3><span class="label label-deviation">Deviation</span> Merged three test files into one main_test.go</h3>
+      <p><strong>Spec:</strong> The AC parenthetical "(及 main_test.go)" implies tests should also collapse to a single file.</p>
+      <p><strong>Implemented:</strong> Merged run_test.go (dispatch seam), config_file_test.go (applyFileConfig precedence), and the residual prompts_config_test.go (applyFileConfig prompts pass-through) into <code>main_test.go</code>, preserving every assertion verbatim. Deleted the three originals.</p>
+      <p><strong>Why:</strong> With all tested symbols now in main.go, splitting the tests across files added no value; one main_test.go matches the AC's "only main.go and main_test.go" shape and keeps the co-located tests discoverable.</p>
+    </div>
+
+    <h2><span class="dot tradeoff"></span> Tradeoffs</h2>
+    <div class="item">
+      <h3><span class="label label-tradeoff">Tradeoff</span> Golden verification by hand-run vs a scripted golden harness</h3>
+      <p><strong>Alternatives:</strong> (a) build a golden-file test comparing full --help/--version output byte-for-byte; (b) run the binary manually and eyeball the key contracts.</p>
+      <p><strong>Pros/cons:</strong> (a) is more durable but --help embeds a registry-derived provider list that drifts as providers are added, making the golden brittle; (b) is quick and confirms the behaviors the AC names (exit codes 0/2/2, --version string, provider help block present).</p>
+      <p><strong>Winner:</strong> (b) — this is a pure code-motion change with no logic edits; the existing dispatch/applyFileConfig unit tests plus manual golden checks fully cover the risk without a brittle new fixture.</p>
+    </div>
+
+    <h2><span class="dot question"></span> Open Questions</h2>
+    <div class="item">
+      <h3><span class="label label-question">Question</span> Is the ~400-line budget a hard ceiling as the CLI grows?</h3>
+      <p><strong>Assumption:</strong> 319 lines leaves headroom; new flags add a few lines each to both the struct and main(). No further extraction is planned.</p>
+      <p><strong>Verify:</strong> If future flag additions push main.go past ~400, revisit whether flag registration should move to a helper (e.g. a registerFlags(opts) function) rather than re-splitting into multiple files.</p>
+    </div>
+
+    <p class="footer">Generated by goal-workflow /note-it</p>
+  </div>
+</body>
+</html>
